handle comment in js when replacing js views
sometime when we are adding comment to js code in views
it doesn't formatted properly and comment being end up in the same line as
previous or next js code line.

This is because of format-couchdb-views is making multiple line js code
to a single string and then performing the code formatting.

Updating replace js code script todo the same, so we don't get un-commit change
in CI.